requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-import-export==2.8.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyYAML==6.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
